Loading from state and fetch ID combined
The state can not be a standalone feature because if a user copies and pastes a link, it causes the application to crash due to it not fetching the database. Therefore, we must first check to see if the problem is in the state, and if not search the database.