Altered addProblems
made visual improvements to the page and the headers now are much easier to distinguish from the input fields. I also made certain inputs able to overflow and change the box size, this is so that the user is able to see their entire inputs rather than a small portion.