User Input for Regex
Allow the user to determine the start and end location when parsing for description, title, and where to split problems. This allows for more accurate use of the PDFParser as it allows the user to account for variances. This will require some basic experience with RegEx on the admin's part, but our current admins are smart enough.