Bug Fixes
Fixed the loading problem with the ChatBox component. This issue would be caused by users loading a chat and then clicking away from the drawer, which would cancel loading.

Fixed the logout issue. This would happen when a user would click logout, but then be able to interact with the website for a few seconds. It now disables everything and shows a loading image.

Finalized the interview problem layout. It will display tags and difficulty when in solution mode.